Update bootstrap-table-cookie.js
The function "cookiesPresent" has a chance of returning "[]" or an empty array when the function "getCookies" returns an empty array itself. This makes the function "cookiesPresent", which SHOULD be a boolean, return an empty array ("[]"). In javascript an empty array "[]" does not equal false. Simple test:

```
var t = [];
if (true && t) {
    console.log("Should never get here!")
}
```

This simple fix checks the length of the array returned from getCookies so that we return a boolean value in said function check instead of returning an array